split out time stuff for reuse.
